Move cells into their own package, and clean up some of the imports
