+++ src/roxene/__init__.py
@@ -1,5 +1,5 @@
-from .cells import Cell, InputCell
-from .neuron import Neuron
 from .organism import Organism, Gene
 from .util import random_neuron_state
 from .persistence import EntityBase
+from .constants import *
+from .cells import Neuron, InputCell
