+++ tests/CreateInputCell_test.py
@@ -4,7 +4,8 @@ from random import Random
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
 
-from roxene import InputCell, Organism, EntityBase
+from roxene import Organism, EntityBase
+from roxene.cells import InputCell
 from roxene.genes import CreateInputCell
 
 SEED = 84592184
